Coerce datums from stages that failed correctly
When an external function invocation fails, the stage is failed with a
datum type of httpresp; similarly, when an externally completable stage
has a failure posted to it, that stage is failed with a datum type of
httpreq.

These are correctly coerced into wrapping exceptions by calls to
waitForCompletion (ie, .get() ) because that queries the graph stage
and receives the stage status together with the datum.

However, those stages are failed, and thus the completer's control flow
will cause the datum to be treated as an error, which will percolate
down through the graph. The only stanges that are triggered at that
point will be error handlers (.exceptionally, .handle, .whenComplete).

These continuations are invoked with a lambda which expects a Throwable;
however, the argument at that point is an {httpreq,httpresp}. We need
to catch that argument and coerce it to an exception before invoking
the lambda.

Unfortunately, direct introspection on the DispatchPattern's returned
method* doesn't provide enough information to correctly know to wrap
these datums.

(* The argument types are all Object)

So, we add additional markers to the closures taken by .exceptionally
et al and use these to trigger the coercion.

Tests are provided to exercise this behaviour.

* nit fix: make the HttpMethod conversion of the completer input
  more robust