Make most of the tests work
Haul in the integration test that exercises as much as it can.

In order for the CloudThreadRuntime to be serialisable, we have to
ensure that the factory that it wraps is also serializable. That means
dropping the direct reference to the InMemoryCompleter. Instead, we
access that through a static member on the FnTestingRule - with the
limitations about running in parallel that that implies.

Tidy up some of the error case handling to return the correctly-wrapped
errors.

For the moment, we redirect the real System.out and System.err for the
duration of the run to our logging stream. All of the logs are available
to a test as a single (potentially interleaved) stream - the idea here
is that the test is looking for a marker *somewhere* in that output,
so we make it all available as a first pass.