async-thumbnails: list files using minio client
Previously we asked users to run `ls` on the minio storage dir which
works on OS X due to the way the volume permissions work however when
running natively on linux you cannot do this as the folder is created
with root access only permissions. We instead use `mc` to list the files
present in the bucket.

Lines starting # with '#' will be ignored, and an empty message aborts
the commit.  # # On branch bugfix/async-thumbnails-minio-permissions #
Changes to be committed: #	modified:   README.md #	modified:
build.sh # modified: setup/setup.sh #