Make it all work
This is complicated. We want to fork classloaders for the runtime and
the api at least, so that we have multiple parallel copies of our
namespaces. However, that means that we need to translate between
universes as we call into our completer and turn the received values
back into ones that conform to our local classloader types.

An option instead of addMirroredClass() is to simply mirror *everything*
by default. The problem here is again, that if a test wants to set a
static value on some external class, then it'll be able to do so,
but the test class won't see that change.

Finally: note that the main event loop is not piped through a forked
classloader. The reason for this was to avoid having to muck around with
reflection in the test cases for non-CT tests. I think that's likely
to be of some use to non-CT testers, so we want to use the default
classloader behaviour for the main function invocation. (cf. the
difference between `new EntryPoint().run` and `fcl.run` in
FnTestingRule.)