Fix up in-mem whenComplete
The semantics of whenComplete are fiddly; in conjunction with the
exception-wrapping mechanics of the completer, getting this right is
similarly fiddly. That's more a consequence of how whenComplete
works, alas.

Include one test for the whenComplete ordering; this test is racy,
unfortunately: it's possible for it to pass on a very loaded machine
even when the problem still exists. However, a correct in-mem
implementation will always pass.