Bring in the pluggable invoker work
Author: Will Price

This work was initially done to support functions written against the
Spring framework. The Spring invoker needs to be landed as a separate
piece.

Support reified paramterised types in coercions

Improve getInvokeInstance and getMethod on RuntimeInvocation javadoc

Add constructor to DefaultMethodWrapper to take a class and string

The JacksonCoercionTests no longer reload the class under test using the
function loader instead just constructing a `DefaultMethodWrapper` in the
test class.