Add a test
See com.fnproject.fn.runtime.PluggableInvokerTest

At the moment, the initial named method (handleRequest in the test)
needs to exist - it is wrapped early in the construction of the
various objects. In the case of an overridden invoker, that method
might not be used.

We should manage something better under those circumstances.