try to add user agent?
I highly doubt this will work. we appear to have some strange incantation of
looking at release.version to release this thing, from some cursory googling
something like this can get the version of the fdk we're using at runtime, but
I am not exactly sure how to do it. I suspect someone fluent in java things
can do this in a few seconds, if I could ask for a little hand here, I think
this is the right place to put it.