Workaround to make the build work after Maven fixed CVE-2021-26291 (#254)
Maven has added a default setting in the settings.xml which blocks access to any HTTP-based repository. However, as part of the FDK build we are running an HTTP server hosting the built JARs in a container and accessing it in a Maven command running in another container (either in the build or through `fn build` in the tests). This is not allowed anymore by default due to Maven's change.
We introduce a workaround that uses a local settings file to temporarily disable the blockage and redirects an HTTP-based repo queries to our local container running the HTTP server.
For the build, the change is only relevant for the copy-dependencies step of the build image creations, and the local settings file is deleted afterwards, in a way that does not even leave a layer with an "insecure" Maven settings file in the container.
For the tests, we have to use a custom Dockerfile template in every test function, which is filled in by the test logic, and we generate a Maven settings file which overrides the standard one for that specific test container.

Also we try to unbreak the CircleCI build by using the latest image and providing a custom Maven settings file with the appropriate blocker and with the fix to SUREFIRE-1588.

Finally, we disable some flakey tests that have to be rewritten anyway.

Co-authored-by: Andrea Rosa <andrea.rosa@oracle.com>