Feature: add support for upcoming OCI Tracing
Oracle Functions will soon release a feature providing an integration with the Oracle Application Performance Monitoring (APM) service. This service supports the sending of Zipkin formatted tracing data to a collector endpoint.

This change adds a TracingContext interface that users can use as a parameter of their function (much like the HTTPGatewayContext) to have access to the configuration data that Zipkin would need (i.e. a collector URL) and the headers data involved with tracing (for Zipkin, the "B3 Headers").

Once the Oracle Functions integration with APM is complete, users will be able to enable tracing for their functions and in that case the TracingContext will report isTracingEnabled() == true and will return the right data for each of its methods.

This change has no effect on existing functions, except that the keys `"OCI_TRACING_ENABLED"` and `"OCI_TRACE_COLLECTOR_URL"` are now reserved for this tracing integration and cannot be used for function configuration.